refactored the code for cloud storage
